move user settings backend feature to new file
