+++ packages/backend/src/index.ts
@@ -5,11 +5,7 @@ import {
   dynamicPluginsFeatureLoader,
   dynamicPluginsFrontendServiceRef,
 } from '@backstage/backend-dynamic-feature-service';
-import {
-  coreServices,
-  createBackendFeatureLoader,
-  createServiceFactory,
-} from '@backstage/backend-plugin-api';
+import { createServiceFactory } from '@backstage/backend-plugin-api';
 import { PackageRoles } from '@backstage/cli-node';
 
 import * as path from 'path';
@@ -21,6 +17,7 @@ import {
   pluginIDProviderService,
   rbacDynamicPluginsProvider,
 } from './modules';
+import { userSettingsBackend } from './modules/userSettings';
 
 // Create a logger to cover logging static initialization tasks
 const staticLogger = WinstonLogger.create({
@@ -166,27 +163,6 @@ backend.add(import('@internal/plugin-dynamic-plugins-info-backend'));
 backend.add(import('@internal/plugin-scalprum-backend'));
 backend.add(import('@internal/plugin-licensed-users-info-backend'));
 
-backend.add(
-  createBackendFeatureLoader({
-    deps: {
-      config: coreServices.rootConfig,
-    },
-    async loader({ config }) {
-      const persistence =
-        config.getOptionalString('userSettings.persistence') ?? 'database'; // default to database
-
-      if (persistence !== 'database' && persistence !== 'browser') {
-        throw new Error(
-          `Invalid config value for 'userSettings.persistence': "${persistence}". Must be either "database" or "browser".`,
-        );
-      }
-      if (persistence === 'database') {
-        return [import('@backstage/plugin-user-settings-backend')];
-      }
-      // Opt-out: browser -> no backend feature
-      return [];
-    },
-  }),
-);
+backend.add(userSettingsBackend);
 
 backend.start();
